+++ app/src/main/java/woowacourse/shopping/data/local/HistoryDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.history
+package woowacourse.shopping.data.local
 
 import androidx.room.Dao
 import androidx.room.Insert
