+++ app/src/main/java/woowacourse/shopping/ui/cart/CartClickListener.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.ui.cart
 
 import woowacourse.shopping.domain.product.Product
-import woowacourse.shopping.ui.QuantityClickListener
+import woowacourse.shopping.utils.QuantityClickListener
 
 interface CartClickListener : QuantityClickListener {
     fun onClick(product: Product)
