+++ app/src/main/java/woowacourse/shopping/ui/fashionlist/ProductClickListener.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.ui.fashionlist
 
 import woowacourse.shopping.domain.product.Product
-import woowacourse.shopping.ui.QuantityClickListener
+import woowacourse.shopping.utils.QuantityClickListener
 
 interface ProductClickListener : QuantityClickListener {
     fun onClick(product: Product)
