+++ app/src/main/java/woowacourse/shopping/utils/QuantityClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.utils
 
 import woowacourse.shopping.domain.product.CartItem
