+++ app/src/main/java/woowacourse/shopping/utils/ViewModelFactory.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.utils
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.domain.product.ProductRepository
-import woowacourse.shopping.ui.viewmodel.ProductListViewModel
+import woowacourse.shopping.ui.fashionlist.ProductListViewModel
 
 object ViewModelFactory {
     fun createProductViewModelFactory(productRepository: ProductRepository): ViewModelProvider.Factory {
