+++ app/src/main/res/layout/activity_product_detail.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="quantityClickListener"
-            type="woowacourse.shopping.ui.QuantityClickListener" />
+            type="woowacourse.shopping.utils.QuantityClickListener" />
 
         <variable
             name="product"
